🔒 Security & Architecture Audit Fixes
✅ Security: Backend Admin Role Validation
- Add AdminGuard with JWT validation for admin endpoints
- Replace frontend-only role checks with backend validation
- Update AdminController to use AdminGuard

⚖️ Driver Queue Fairness
- Fix new drivers going to front of queue
- New drivers now go to end with Date.now() fallback

💰 Negative Balance Protection
- Add balance check in acceptRide (-500₸ threshold)
- Update driver filtering to match threshold

🔄 Ride Retry Logic
- Enhanced timeout + retry mechanism
- Add excluded drivers tracking to prevent duplicate offers
- Max 3 attempts with 30-second timeout

📍 GPS Database Optimization
- Batch updates every 30 seconds already implemented
- Add cache cleanup for stale entries
- Immediate updates for active rides

💾 File Upload Memory Fix
- Already using diskStorage instead of memoryStorage
- Add temp directory creation and validation
- Improve file size and type validation

🔌 WebSocket Memory Leaks
- Proper event listener cleanup with socket.off()
- Fix async/await issues in useEffect
- Add cleanup function storage

🎯 Frontend Improvements
- Fix GraphQL Float types for coordinates
- Add validation for addresses and coordinates
- Fix chat role logic based on auth.role
- Add proper error handling

+++ mobile/src/api/queries.ts
@@ -103,10 +103,10 @@ export const CREATE_RIDE = gql`
     $passengerId: String!
     $fromAddress: String!
     $toAddress: String!
-    $fromLat: Int
-    $fromLng: Int
-    $toLat: Int
-    $toLng: Int
+    $fromLat: Float
+    $fromLng: Float
+    $toLat: Float
+    $toLng: Float
     $paymentMethod: String
     $stops: [StopInput]
   ) {
